{{cookiecutter.project_slug}}/requirements.txt
# Core dependencies
pydantic==2.5.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# AI/ML
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0

# Database
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# Utilities
config-box
langchain-mongodb

# Logging
# prometheus_client>=0.17.1
# watchtower

# Web framework
streamlit
fastapi
uvicorn

# Authentication
pyJWT
requests==2.32.5


